objtool: Make recordmcount into an objtool subcmd
Rather than a standalone executable merge recordmcount as a sub
command of objtool. This is a small step towards cleaning up
recordmcount and eventually saving ELF code with objtool.

For the initial step all that's required is a bit of Makefile
changes and invoking the former main() function from recordmcount.c
because the subcommand code uses similar function arguments as
main when dispatching.

Subsequent patches will gradually convert recordmcount to use
more and more of objtool's code. This will both clean up recordmcount
and share code.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>