objtool: Add a tool to produce a sequence of config changes
This tool will be useful for testing multiple combinations of configs.
Change configs such that:

1) We only change one value at a time. This makes reasoning about
   successive runs easier. NOTE: Use config bisect if you already
   have a bug to hunt.

2) We change a different configuration at each iteration. This makes
   broader variation in configuration more likely over a smaller
   number of runs and hopefully will find bugs sooner than a simpler
   combinatorial walk might. Once we've changed each variable once
   we can return to the first changed variable again.

Use -r [SEED] option to randomly order the selected values. Supply a
SEED to get a reproducible sequence of values.

Use -s to specify the number of "samples" to test. Since we're talking
about combinatorial testing this is useful to try and get a heuristic
sample of sufficiently-different configs without sampling large fractions
of the combinatiorial space.

On input the tool consumes a file describing which variables to change
and which values we'd like to try. An empty set of values implies a
tristate (yes, no, or module) configuration variable:

./config-sample - <<-EOF
CONFIG_FOO
CONFIG_BAR valA valB
EOF

Produces:

CONFIG_FOO=y
CONFIG_BAR=valA
CONFIG_FOO=n
CONFIG_BAR=valB
CONFIG_FOO=m
CINFIG_BAR=valA

Each line represents the change to make to an existing config.

Note the tool does NOT:

	1) Check that these variables exist
	2) Much less what their values are in the "base" config to modify
	3) Nor that the proposed values are "legal"

This is also not a config patching or bisecting tool. It's not a
patching tool because of point #2 above because the value of the
config variable in the base config is not necessarily in the set of
values supplied. It's up to the user to select an appropriate
base config, the variables, and their values.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>