objtool: Add a test script for builds invoking recordmcount
Use config-sample to test several variations on a base config
and ensure that combinations of ftrace, various HAVE_, and
other configs compile and produce the expected results.

For now we keep the testing simple and only test on HAVE_C_RECORDMCOUNT
This allows faster testing cycles for each subsequent patch
that merges recordmcount into objtool.

The goal here is to ensure that the same mcount sections are output --
We selected the CONFIG_* variables such that the sections should not
change what gets included. Then we use readelf and extract_kernel to
fetch the recordmcount data from the built kernel and diff against the
previous iteration to ensure it does not change.

Note that it doesn't make sense to always run this test because not
all architectures set HAVE_C_RECORDMCOUNT. So this is not a candidate
for validating kernels.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>