objtool: Check recordmcount output for changes
Use config-sample to test several variations on a base config
and ensure that combinations of ftrace, various HAVE_, and
other configs compile and we can compare the mcount sections
produced.

We selected the CONFIG_* variables such that the sections should not
change what gets included. Then we use readelf and extract_kernel to
fetch the recordmcount data from the built kernel and diff against the
previous iteration to ensure it does not change.

This allows us to rerun the test script on each subsequent commit
to verify that the mcount data does not change even as we convert
more and more of recordmcount.[ch] to use libelf/objtool accessors.

For now we keep the testing simple and only test on HAVE_C_RECORDMCOUNT
This allows faster testing cycles for each subsequent patch
that merges recordmcount into objtool.

Note that it doesn't make sense to always run this test because not
all architectures set HAVE_C_RECORDMCOUNT. So this is not a candidate
for general validation of built kernels.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>