objtool: recordmcount: Use libelf section header name
Rather than passing in the string table contents as a parameter,
pass in the section index of the string table and rely on elf
string table accessor functions look up section names. Once
we're only using the libelf structure types we'll be able to
drop the giant #ifdef block at the top of recordmcount.h and
simplify Elf 32/64-bit support.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>