objtool: recordmcount: Make __has_rel_mcount() 32/64-bit
Pass section pointers around instead of ELF header pointers. This
eliminates a lot of redundant indexing and simplifies handling of
things like section names.

While we're here, inline __has_rel_mcount(). This allows us to remove
a comment about ISO C that no longer seems relevant and makes the
comment about assuming the section pointer is a rel[a] section
irrelevant.

The more interesting part here is we need to mark a section we're
modifying as "changed" using the right objtool field and function.
We need to do this because we're no longer relying on the section
header of the section name string table pointing directly to an entry
in the section table.

That said, we're still appending the new sections without going
through the objtool code. This is a bit of a hack we'll remove
in subsequent patches.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>