objtool: mcount: Pre-allocate new ELF sections
Rather than allocating the ELF sections after collecting the
mcount locations and building the relocation entries, create
the empty sections beforehand. This has the benefit of
removing the memcpy() and just using the resulting buffers
directly and it makes do_func()'s control flow even clearer
so we can remove the result variable.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>