vsock: Enable selecting transport via setsockopt()
Currently we allow new code to use a new socket option to set the
transport. Alternative solutions might exist which haven't been
explored:
    Use netlink somehow?
    Add ioctls for adjusting the underlying transport

NOTES and TODOs:
Separate from the other patches in case the RFC suggests a better
method of selecting the transport.

See "Implement" and "Review" TODOs below.

Actually getting and selecting a transport with getsockopt and
setsockopt and have not been tested.

Review:

Review/fix socket lock probably needs to be taken to change transport in
    <setsocktopt> calls to __vsock_set_transport()

Signed-off-by: Matt Helsley <mhelsley@vmware.com>