objtool: check: Handle overwriting current CFA base register
When overwriting the current call frame base register with a value
that doesn't have specific semantics for objtool, try to keep track
of the call frame with the initial call frame base register. This
can only be done when the initial base register is SP as it's
progression is tracked in insn_state->stack_size.

Signed-off-by: Julien Thierry <jthierry@redhat.com>