Extend mcount checking -- roughly diff objdump output
When called with the output from one objdump the awk script
will check it. Two or more causes the script to check each
and then compare to the first for any inconsistencies. In
this way we could use multiple tools or multiple revisions
of tools to produce the mcount data and then we can compare
them -- all at once if necessary -- to see if they're
"correct" and what, if anything, changes.

Signed-off-by: Matt Helsley <mhelsley@vmware.com>